% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs-binom-pois.R
\name{DIC for Metropolis-within-Gibbs Sampler}
\alias{DIC for Metropolis-within-Gibbs Sampler}
\alias{DIC.gibbs.mixlink.reg}
\title{DIC for Metropolis-within-Gibbs}
\usage{
DIC.gibbs.mixlink.reg(gibbs.out, y, X, trials = NULL, offset = NULL,
  invlink = NULL, family = NULL)
}
\arguments{
\item{gibbs.out}{Results from Metropolis-within-Gibbs sampler.}

\item{y}{Observations.}

\item{X}{Design matrix for regression.}

\item{trials}{Number of success/failure trials.}

\item{offset}{Offset term to add to regression function, as commonly used in count
models.}

\item{invlink}{The inverse link function for the mean. Default is \code{NULL},
which indicates to use \code{plogis} for Binomial and \code{exp} for Poisson.}

\item{family}{Can be either \code{binomial} or \code{poisson}.}
}
\value{
DIC
}
\description{
Deviance Information Criteria (DIC) computed on results from
Metropolis-within-Gibbs sampler.
}

