% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{DIC}
\alias{DIC}
\title{Deviance Information Criteria for MCMC Draws}
\usage{
DIC(object, ...)
}
\arguments{
\item{object}{Results from sampler.}

\item{...}{optional arguments (currently not used).}
}
\value{
DIC
}
\description{
Deviance Information Criteria (DIC) computed on results from
MCMC sampler.
}
