% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mlogit}
\alias{mlogit}
\alias{mlogit}
\alias{inv.mlogit}
\title{Multinomial logit transformation}
\usage{
mlogit(p)

inv.mlogit(x)
}
\arguments{
\item{p}{A J-dimensional numeric vector in the probabiity simplex}

\item{x}{A (J-1)-dimensional numeric vector}
}
\value{
A numeric vector
}
\description{
mlogit transforms from probability simplex S^J to R^(J-1)
inv.mlogit transforms from R^(J-1) to probability simplex S^J
}
\examples{
  x <- c(1,1,1,1)
  p <- inv.mlogit(x)
  x2 <- mlogit(p)
}
