% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
\code{x / sum(x)}
}
\description{
Scale a numeric vector by the sum of its elements
}
\examples{
  x <- c(1,1,1,1)
  normalize(x)
}
