% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs-binom-pois.R, R/mixlink-package.R
\name{S3 methods for gibbs.mixlink output objects}
\alias{S3 methods for gibbs.mixlink output objects}
\alias{summary.gibbs.mixlink}
\alias{S3 methods for gibbs.mixlink output objects}
\alias{print.gibbs.mixlink}
\alias{S3 methods for gibbs.mixlink output objects}
\alias{DIC.gibbs.mixlink}
\alias{S3 methods for gibbs.mixlink output objects}
\alias{residuals.gibbs.mixlink}
\alias{S3 methods for gibbs.mixlink output objects}
\title{S3 methods for gibbs.mixlink output objects}
\usage{
\method{summary}{gibbs.mixlink}(object, ...)

\method{print}{gibbs.mixlink}(x, ...)

\method{DIC}{gibbs.mixlink}(object, ...)

\method{residuals}{gibbs.mixlink}(object, ...)
}
\arguments{
\item{object}{An \code{gibbs.mixlink} object from \code{gibbs.mixlink}.}

\item{...}{Not currently used.}

\item{x}{Same argument type as \code{object}.}
}
\description{
S3 methods for gibbs.mixlink output objects
}
\details{
\code{DIC} returns the Deviance Information Criterion computed from MCMC draws.
  \code{residuals} returns randomized quantile residuals averaged over MCMC draws.
}
