% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs-binom-pois.R, R/mixlink-package.R
\name{S3 methods for gibbs.mixlink.reg output objects}
\alias{S3 methods for gibbs.mixlink.reg output objects}
\alias{print.gibbs.mixlink.reg}
\alias{summary.gibbs.mixlink.reg}
\title{S3 methods for gibbs.mixlink.reg output objects}
\usage{
\method{summary}{gibbs.mixlink.reg}(object, ...)

\method{print}{gibbs.mixlink.reg}(x, ...)
}
\arguments{
\item{object}{An \code{gibbs.mixlink.reg} object from \code{gibbs.mixlink.reg}.}

\item{...}{Not currently used.}

\item{x}{Same argument type as \code{object}.}
}
\description{
S3 methods for gibbs.mixlink.reg output objects
}

