% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-mle.R
\name{S3 methods for mle.fit.ci objects}
\alias{S3 methods for mle.fit.ci objects}
\alias{print.mle.fit.ci}
\title{Print summary of an \code{mle.fit.ci} object}
\usage{
\method{print}{mle.fit.ci}(x, ...)
}
\arguments{
\item{x}{An \code{mle.fit.ci} object from \code{confint.fit.mle}.}

\item{...}{Not currently used.}
}
\description{
See generic \code{print} function for usage.
}
