% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrop-binom.R, R/metrop-pois.R,
%   R/mixlink-package.R
\name{S3 methods for rwmetrop output objects}
\alias{S3 methods for rwmetrop output objects}
\alias{summary.rwmetrop.mixlink.binomial}
\alias{S3 methods for rwmetrop output objects}
\alias{print.rwmetrop.mixlink.binomial}
\alias{S3 methods for rwmetrop output objects}
\alias{DIC.rwmetrop.mixlink.binomial}
\alias{S3 methods for rwmetrop output objects}
\alias{residuals.rwmetrop.mixlink.binomial}
\alias{S3 methods for rwmetrop output objects}
\alias{summary.rwmetrop.mixlink.poisson}
\alias{S3 methods for rwmetrop output objects}
\alias{print.rwmetrop.mixlink.poisson}
\alias{S3 methods for rwmetrop output objects}
\alias{DIC.rwmetrop.mixlink.poisson}
\alias{S3 methods for rwmetrop output objects}
\alias{residuals.rwmetrop.mixlink.poisson}
\alias{S3 methods for rwmetrop output objects}
\title{S3 methods for rwmetrop output objects}
\usage{
\method{summary}{rwmetrop.mixlink.binomial}(object, ...)

\method{print}{rwmetrop.mixlink.binomial}(x, ...)

\method{DIC}{rwmetrop.mixlink.binomial}(object, ...)

\method{residuals}{rwmetrop.mixlink.binomial}(object, ...)

\method{summary}{rwmetrop.mixlink.poisson}(object, ...)

\method{print}{rwmetrop.mixlink.poisson}(x, ...)

\method{DIC}{rwmetrop.mixlink.poisson}(object, ...)

\method{residuals}{rwmetrop.mixlink.poisson}(object, ...)
}
\arguments{
\item{object}{An \code{rwmetrop.mixlink.binomial} object from \code{rwmetrop.mixlink.binomial}
or an \code{rwmetrop.mixlink.poisson} object from \code{rwmetrop.mixlink.poisson}.}

\item{...}{Not currently used.}

\item{x}{Same argument type as \code{object}.}
}
\description{
S3 methods for rwmetrop output objects
}
\details{
\code{DIC} returns the Deviance Information Criterion computed from MCMC draws.
  \code{residuals} returns randomized quantile residuals averaged over MCMC draws.
}
