% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-mle.R, R/mixlink-package.R
\name{S3 methods for mle.fit objects}
\alias{AIC.mle.fit}
\alias{BIC.mle.fit}
\alias{S3 methods for mle.fit objects}
\alias{coef.mle.fit}
\alias{confint.mle.fit}
\alias{logLik.mle.fit}
\alias{print.mle.fit}
\alias{summary.mle.fit}
\alias{vcov.mle.fit}
\title{S3 methods for mle.fit objects}
\usage{
\method{confint}{mle.fit}(object, parm = NULL, level = 0.95, ...)

\method{print}{mle.fit}(x, ...)

\method{summary}{mle.fit}(object, ...)

\method{coef}{mle.fit}(object, ...)

\method{logLik}{mle.fit}(object, ...)

\method{AIC}{mle.fit}(object, ..., k = 2)

\method{BIC}{mle.fit}(object, ...)

\method{vcov}{mle.fit}(object, ...)
}
\arguments{
\item{object}{An \code{mle.fit} object from \code{fit.mle}.}

\item{parm}{Not currently used.}

\item{level}{Desired confidence level.}

\item{...}{Not currently used.}

\item{x}{An \code{mle.fit} object from \code{fit.mle}.}

\item{k}{Penalty term is \eqn{k q} where \eqn{q =\dim \bm{\theta}}. \eqn{k=2} is the
default for AIC calculation.}
}
\description{
S3 methods for mle.fit objects
}

