% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrop-binom.R, R/metrop-pois.R, R/mixlink-package.R
\name{S3 methods for rwmetrop output objects}
\alias{S3 methods for rwmetrop output objects}
\alias{print.rwmetrop.mixlink.binomial}
\alias{print.rwmetrop.mixlink.poisson}
\alias{summary.rwmetrop.mixlink.binomial}
\alias{summary.rwmetrop.mixlink.poisson}
\title{S3 methods for rwmetrop output objects}
\usage{
\method{summary}{rwmetrop.mixlink.binomial}(object, ...)

\method{print}{rwmetrop.mixlink.binomial}(x, ...)

\method{summary}{rwmetrop.mixlink.poisson}(object, ...)

\method{print}{rwmetrop.mixlink.poisson}(x, ...)
}
\arguments{
\item{object}{An \code{rwmetrop.mixlink.binomial} object from \code{rwmetrop.mixlink.binomial}
or an \code{rwmetrop.mixlink.poisson} object from \code{rwmetrop.mixlink.poisson}.}

\item{...}{Not currently used.}

\item{x}{Same argument type as \code{object}.}
}
\description{
S3 methods for rwmetrop output objects
}

